factorization into primes along with tests
